[ESKL-83/resume-match] - Removes unnecessary boilerplate code
