+++ 2048/2048/src/main/java/com/appcoins/eskills2048/model/UserDetailsHelper.java
@@ -1,23 +1,29 @@
 package com.appcoins.eskills2048.model;
 
 import java.util.List;
+import javax.inject.Inject;
+import javax.inject.Singleton;
 
+@Singleton
 public class UserDetailsHelper {
-    private User currentUser;
+  private User currentUser;
 
-    public User getNextOpponent(List<User> opponents) {
-        if (opponents.size() == 1 || currentUser == null) {
-            currentUser = opponents.get(0);
-            return currentUser;
-        }
+  @Inject public UserDetailsHelper() {
+  }
 
-        for (User opponent : opponents) {
-            if (!opponent.getWalletAddress().equalsIgnoreCase(
-                    currentUser.getWalletAddress())) {
-                currentUser = opponent;
-                break;
-            }
-        }
-        return currentUser;
+  public User getNextOpponent(List<User> opponents) {
+    if (opponents.size() == 1 || currentUser == null) {
+      currentUser = opponents.get(0);
+      return currentUser;
     }
+
+    for (User opponent : opponents) {
+      if (!opponent.getWalletAddress()
+          .equalsIgnoreCase(currentUser.getWalletAddress())) {
+        currentUser = opponent;
+        break;
+      }
+    }
+    return currentUser;
+  }
 }
