+++ 2048/2048/src/main/java/com/appcoins/eskills2048/repository/LocalGameStatusRepository.java
@@ -4,7 +4,10 @@ import com.appcoins.eskills2048.Tile;
 import com.appcoins.eskills2048.model.LocalGameStatus;
 import com.appcoins.eskills2048.util.GameFieldConverter;
 import com.appcoins.eskills2048.util.UserDataStorage;
+import javax.inject.Inject;
+import javax.inject.Singleton;
 
+@Singleton
 public class LocalGameStatusRepository {
   private static final String GAME_STATUS_PREFIX = "PREFERENCES_GAME_STATUS_";
   private static final String SESSION = "SESSION";
@@ -15,7 +18,7 @@ public class LocalGameStatusRepository {
   private final UserDataStorage userDataStorage;
   private final GameFieldConverter gameFieldConverter;
 
-  public LocalGameStatusRepository(UserDataStorage userDataStorage,
+  @Inject public LocalGameStatusRepository(UserDataStorage userDataStorage,
       GameFieldConverter gameFieldConverter) {
     this.userDataStorage = userDataStorage;
     this.gameFieldConverter = gameFieldConverter;
