+++ 2048/2048/src/main/java/com/appcoins/eskills2048/repository/RoomRepository.java
@@ -4,20 +4,18 @@ import com.appcoins.eskills2048.api.RoomApi;
 import com.appcoins.eskills2048.model.PatchRoomRequest;
 import com.appcoins.eskills2048.model.RoomResponse;
 import com.appcoins.eskills2048.model.UserStatus;
-import io.reactivex.Flowable;
 import io.reactivex.Single;
-import java.net.SocketTimeoutException;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicInteger;
-import retrofit2.HttpException;
+import javax.inject.Inject;
+import javax.inject.Singleton;
 
+@Singleton
 public class RoomRepository {
 
   public static final String BEARER_ = "Bearer ";
 
   private final RoomApi roomApi;
 
-  public RoomRepository(RoomApi roomApi) {
+  @Inject public RoomRepository(RoomApi roomApi) {
     this.roomApi = roomApi;
   }
