+++ 2048/2048/src/main/java/com/appcoins/eskills2048/repository/StatisticsRepository.java
@@ -3,11 +3,14 @@ package com.appcoins.eskills2048.repository;
 import com.appcoins.eskills2048.api.GeneralPlayerStats;
 import com.appcoins.eskills2048.model.GeneralPlayerStatsResponse;
 import io.reactivex.Single;
+import javax.inject.Inject;
+import javax.inject.Singleton;
 
+@Singleton
 public class StatisticsRepository {
   private final GeneralPlayerStats api;
 
-  public StatisticsRepository(GeneralPlayerStats api) {
+  @Inject public StatisticsRepository(GeneralPlayerStats api) {
     this.api = api;
   }
