+++ 2048/2048/src/main/java/com/appcoins/eskills2048/usecase/GetRoomUseCase.java
@@ -4,12 +4,15 @@ import com.appcoins.eskills2048.model.RoomResponse;
 import com.appcoins.eskills2048.repository.RoomRepository;
 import io.reactivex.Single;
 import io.reactivex.schedulers.Schedulers;
+import javax.inject.Inject;
+import javax.inject.Singleton;
 
+@Singleton
 public class GetRoomUseCase {
 
   private final RoomRepository roomRepository;
 
-  public GetRoomUseCase(RoomRepository roomRepository) {
+  @Inject public GetRoomUseCase(RoomRepository roomRepository) {
     this.roomRepository = roomRepository;
   }
