+++ 2048/2048/src/main/java/com/appcoins/eskills2048/usecase/GetUserStatisticsUseCase.java
@@ -5,11 +5,14 @@ import com.appcoins.eskills2048.repository.StatisticsRepository;
 import com.appcoins.eskills2048.repository.StatisticsTimeFrame;
 import io.reactivex.Single;
 import io.reactivex.schedulers.Schedulers;
+import javax.inject.Inject;
+import javax.inject.Singleton;
 
+@Singleton
 public class GetUserStatisticsUseCase {
   private final StatisticsRepository statisticsRepository;
 
-  public GetUserStatisticsUseCase(StatisticsRepository statisticsRepository) {
+  @Inject public GetUserStatisticsUseCase(StatisticsRepository statisticsRepository) {
     this.statisticsRepository = statisticsRepository;
   }
