+++ 2048/2048/src/main/java/com/appcoins/eskills2048/usecase/SetScoreUseCase.java
@@ -3,19 +3,17 @@ package com.appcoins.eskills2048.usecase;
 import com.appcoins.eskills2048.model.RoomResponse;
 import com.appcoins.eskills2048.model.UserStatus;
 import com.appcoins.eskills2048.repository.RoomRepository;
-import io.reactivex.Flowable;
 import io.reactivex.Single;
 import io.reactivex.schedulers.Schedulers;
-import java.net.SocketTimeoutException;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicInteger;
-import retrofit2.HttpException;
+import javax.inject.Inject;
+import javax.inject.Singleton;
 
+@Singleton
 public class SetScoreUseCase {
 
   private final RoomRepository roomRepository;
 
-  public SetScoreUseCase(RoomRepository roomRepository) {
+  @Inject public SetScoreUseCase(RoomRepository roomRepository) {
     this.roomRepository = roomRepository;
   }
