+++ 2048/2048/src/main/java/com/appcoins/eskills2048/util/GameFieldConverter.java
@@ -2,19 +2,22 @@ package com.appcoins.eskills2048.util;
 
 import com.appcoins.eskills2048.Tile;
 import com.google.gson.Gson;
+import javax.inject.Inject;
+import javax.inject.Singleton;
 
+@Singleton
 public class GameFieldConverter {
-    private final Gson gson;
+  private final Gson gson;
 
-    public GameFieldConverter(Gson gson) {
-        this.gson = gson;
-    }
+  @Inject public GameFieldConverter(Gson gson) {
+    this.gson = gson;
+  }
 
-    public String fromField(Tile[][] field) {
-        return gson.toJson(field);
-    }
+  public String fromField(Tile[][] field) {
+    return gson.toJson(field);
+  }
 
-    public Tile[][] toField(String field) {
-        return gson.fromJson(field, Tile[][].class);
-    }
+  public Tile[][] toField(String field) {
+    return gson.fromJson(field, Tile[][].class);
+  }
 }
